BARK-87: Move SubProfile into the dog-profile-models module
