+++ src/app/user/(logged-in)/my-pets/edit-dog/[dogId]/page.tsx
@@ -8,7 +8,10 @@ import { EditDogProfileFormController } from "../../_lib/components/edit-dog-pro
 import { SimpleErrorPage } from "@/app/_components/simple-error-page";
 import { SubProfileFormController } from "../../_lib/components/sub-profile-form-controller";
 import { opFetchDogReportCount } from "@/lib/bark/operations/op-fetch-dog-report-count";
-import { SubProfile, SubProfileSchema } from "@/lib/bark/models/sub-profile";
+import {
+  SubProfile,
+  SubProfileSchema,
+} from "@/lib/bark/models/dog-profile-models";
 import { Err, Ok, Result } from "@/lib/utilities/result";
 import { DogProfile } from "@/lib/bark/models/dog-profile-models";
 import { CODE } from "@/lib/utilities/bark-code";
